Code tidy-up
Separated related code blocks into their own files

Renamed 'kontent' and 'parsing' for better clarity
